refactoring submit_scans.sh to include more parameters for detect, make the loop contingent on the number of scans; refactoring download script to optionally use a locally populated jars directory for building the docker image from

+++ src/hub_load/download-packages.sh
@@ -5,9 +5,19 @@
 WORKDIR=$(dirname $0)
 mkdir -p $WORKDIR/jars
 
-echo "Downloading and unpacking jars.zip from S3"
-wget https://bds-sa-data-files.s3.us-east-2.amazonaws.com/jars.zip
-unzip jars.zip
-echo "Removing jars.zip now that we have unzipped it"
-rm jars.zip
+#
+# if the jars directory already exists it is assumed that you have created and populated the jars
+# directory locally (not in git) with a bunch of jars that will be used when creating the docker image
+# by virtue of the ADD directive which will add the jars directory to the image
+#
+# Otherwise, this script will pull down a jars.zip from S3 and unpack it to populate the jars directory
+# on the image by running this script in the docker build
+#
+if [ ! -d jars ]; then
+	echo "Downloading and unpacking jars.zip from S3"
+	wget https://bds-sa-data-files.s3.us-east-2.amazonaws.com/jars.zip
+	unzip jars.zip
+	echo "Removing jars.zip now that we have unzipped it"
+	rm jars.zip
+fi
